+++ apps/desktop/src-tauri/src/app/extensions.rs
@@ -1,6 +1,5 @@
 use super::conf;
 use crate::fc::exists;
-use download_npm;
 use serde::{Deserialize, Serialize};
 use std::{fs::create_dir, path::PathBuf, vec};
 
@@ -60,8 +59,6 @@ impl AppExtensions {
             let _ = create_dir(Self::dir_path());
         }
 
-        // let _ = &Self::downloadBuiltInExtensions(self.clone()).await;
-
         // load extensions
         let mut extensions = vec![];
         let dir = Self::dir_path();
@@ -82,24 +79,6 @@ impl AppExtensions {
         }
 
         self.extensions = extensions;
-
-        self
-    }
-
-    #[allow(non_snake_case)]
-    pub async fn downloadBuiltInExtensions(self) -> Self {
-        if exists(&Self::dir_path().join("mdviewy-theme-template")) {
-            return self;
-        }
-
-        let _ = download_npm::download(
-            "mdviewy-theme-template",
-            download_npm::DownloadOptions {
-                untar: true,
-                dest_path: Self::dir_path().to_str().unwrap().to_string(),
-            },
-        )
-        .await;
 
         self
     }